requirements.txt
fastapi
uvicorn[standard]
google-generativeai
PyPDF2
beautifulsoup4
lxml
streamlit

